\name{getformula}
\alias{getformula}
\title{Retrieve a formula object from a model object}
\description{
Retrieve a formula object from a model object.
}
\usage{
getformula(x)
}
\details{
The generic \code{getformula} function serves to extract a formula from a model object.
}
\arguments{
\item{x}{ A network model. }
}
